+++ src/Blocks/components/error/error.php
@@ -20,7 +20,7 @@ $errorValue = Helpers::checkAttr('errorValue', $attributes, $manifest);
 $errorId = Helpers::checkAttr('errorId', $attributes, $manifest);
 $errorTwSelectorsData = Helpers::checkAttr('errorTwSelectorsData', $attributes, $manifest);
 
-$twClasses = FormsHelper::getTwSelectors($errorTwSelectorsData, ['error'], $attributes);
+$twClasses = FormsHelper::getTwSelectors($errorTwSelectorsData, ['error']);
 
 $errorClass = Helpers::classnames([
 	FormsHelper::getTwBase($twClasses, 'error', $componentClass),